Replaces contents of posenet.js
Originally, test-posenet.js was supposed to be just a test, but it ended
up becoming the main file (sorry for that).

This commit replaces the contents of posenet.js with the contents of
test-posenet.js. It also removes test-posenet.js and updates the launch
file.
